cli: deflake Example_nodelocal
The Example_nodelocal spec was failing when the test was run in
parallel. This was due to a race to one test deletingthe file that was
needed by another parallel run of the test. This change creates a unique
name for each test run to avoid this problem.

Release note: None