backupccl: fix temp db ID conflict in cluster restore
Previously, when computing the ID of the temporary system database,
restore did not consider non-table IDs and could produce a conflicting
ID. This commit ensures that all descriptors in the backup are
considered when computing the largest used descriptor ID.

Release note (bug fix): On cluster backups where the largest ID was a
database (typically seen when the last thing the user created was a
database), cluster restore would produce an error. It does not anymore.