backupccl: fix flake in backup stats tests
This commit updates the way we generate table stats in tests to ensure
that the stats cache always sees the values we insert. This change takes
advantage of the fact that inject stats directly to a table will
invalidate the stats cache, so when BACKUP queries the stats it will
find the latest information.

Release note: None