backupccl: only re-create schema change jobs on non-cluster restore
This commit prevents cluster restore from re-synthesizing schema change
jobs operating on the descriptors. This is desirable since we already
restore the original jobs as part of the cluster restore.

Release note (bug fix): When performing a cluster restore of a backup
that had tables which were experiencing a schema change, any in progress
schema changes would be duplicated.