backupccl: refactor split and scatter processor to use processor base
This commit refactors the split and scatter processor to use
ProcessorBase.  Although this processor doesn't have any inputs and the
main utility that ProcessorBase provides is management of draining
inputs, there is still some utility for this processor to embed a
ProcessorBase. As tracing and other infrastructural improvements are
made to ProcessorBase, this allows the processor to stay up to date with
these changes.

The way that this processor works is that it kicks off a goroutine on
start that will begin the split and scatter processor and start sending
progress updates over an internally maintained channel. Next() will
continually read from that channel until it is fully consumed, and then
return any errors that may have been returned if the channel closed
early.

Release note: None