backupccl: add flag to write files in backup processor
Previously, the ExportRequest that backup issues would be responsible
for writing out the SST files to the specified locality. When
considering multi-tenant clusters, since tenants share the same KV
layer it is preferrable for the file-writing to happen in in the
processor itself. This enables controlling the connection to external
resources and is sometimes required (e.g. userfile destinations).

A note on testing: this commit randomly enables the flag which moves the
writing from the KV layer to the SQL layer. Manual testing was done to
exercise that the flag actually works. I expect more testing to be added
when tenant backups are fully integrated and we can test performing a
tenant backup to userfile destinations.

Release note: None