backupccl: group relevant data specification in restoreDataBundle
Rather than passing several related structs around, this commit groups
re-keying, pkIDs, and the spans that are to be restored in one struct.

This allows mutation of the data that is to be restored to be defined in
a single place on this struct.

There should be no changes in functionality.

Release note: None