streamingest: add time offset to stream ingestion
The stream ingestion processor can now ingest a stream of events and
change the timestamps of the events with some fixed offset.

This can be used when considering the following use cases:
- Ingesting a stream that was separated temporily (e.g. a stream created
by the same cluster, but in the past). In this case, we'll want to
consume the events at a new timestamp.
- Streaming after restoring from a backup. When restoring a backup, all
keys in the backup are written at the timestamp of the backup. Stream
events ingested on top of this backup need to be offset based on the
time difference between the BACKUP timestamp and the RESTORE timestamp
of that backup.

Release note: None